ShopExample.Services/PostCategoryService.cs
using ShopExample.Data.Infrastructure;
using ShopExample.Data.Repositories;
using ShopExample.Model.Model;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ShopExample.Services
{
    public interface IPostCategoryService
    {
        PostCategory Add(PostCategory pc);
        PostCategory Delete(long id);
        void Update(PostCategory pc);
        IEnumerable<PostCategory> GetAll();
        IEnumerable<PostCategory> GetAllByParentId(long parentId);
        PostCategory GetById(long id);
    }

    public class PostCategoryService : IPostCategoryService
    {
        IPostCategoryRepository _postCategoryRepository;
        IUnitOfWork _unitOfWork;

        public PostCategoryService(IPostCategoryRepository postCategoryRepository, IUnitOfWork unitOfWork)
        {
            this._postCategoryRepository = postCategoryRepository;
            this._unitOfWork = unitOfWork;
        }

        public PostCategory Add(PostCategory pc)
        {
            return _postCategoryRepository.Add(pc);
        }

        public PostCategory Delete(long id)
        {
            return _postCategoryRepository.Delete(id);
        }

        public IEnumerable<PostCategory> GetAll()
        {
            return _postCategoryRepository.GetAll();
        }

        public IEnumerable<PostCategory> GetAllByParentId(long parentId)
        {
            return _postCategoryRepository.GetMulti(x => x.Status && x.ParentID == parentId);
        }

        public PostCategory GetById(long id)
        {
            return _postCategoryRepository.GetSingleById(id);
        }

        public void Update(PostCategory pc)
        {
            _postCategoryRepository.Update(pc);
        }
    }
}
